Create GameHandler for handling gamestate
GameHandler is a class that manages the initialization and change of
state for a specific instance of a Minesweeper game. It is meant to
take in the input (a user action, which will be some kind of click on
a certain coordinate) and will output the next game state resulting from
that action (win, lose, opened more squares, flagged some squares, etc.).